+++ fennecs/World.CRUD.cs
@@ -30,33 +30,7 @@ public partial class World
         newArchetype.BackFill(typeExpression, data, 1);
     }
 
-
-    internal void AddComponent<T>(Entity3 entity, TypeExpression typeExpression, T data) where T : notnull
-    {
-        if (data == null) throw new ArgumentNullException(nameof(data));
-
-        if (Mode == WorldMode.Deferred)
-        {
-            _deferredOperations.Enqueue(new DeferredOperation {Opcode = Opcode.Add, Identity = entity, TypeExpression = typeExpression, Data = data});
-            return;
-        }
-        
-        AssertAlive(entity);
-
-        ref var meta = ref _meta[entity.index];
-        var oldArchetype = meta.Archetype;
-
-        if (oldArchetype.Signature.Matches(typeExpression)) throw new ArgumentException($"Entity {entity} already has a component of type {typeExpression}");
-
-        var newSignature = oldArchetype.Signature.Add(typeExpression);
-        var newArchetype = GetArchetype(newSignature);
-        Archetype.MoveEntry(meta.Row, oldArchetype, newArchetype);
-
-        // Back-fill the new value
-        newArchetype.BackFill(typeExpression, data, 1);
-    }
-
-
+    
     internal void RemoveComponent(Identity identity, TypeExpression typeExpression)
     {
         if (Mode == WorldMode.Deferred)